Add firs load logick
replase use effect in layout
