+++ src/App.jsx
@@ -1,6 +1,5 @@
-import { lazy, useEffect } from 'react';
-import { Navigate, Route, Routes, useNavigate } from 'react-router-dom';
-import { toast } from 'react-toastify';
+import { lazy } from 'react';
+import { Navigate, Route, Routes } from 'react-router-dom';
 
 import './styles/variables.css';
 import 'modern-normalize/modern-normalize.css';
@@ -8,13 +7,6 @@ import 'modern-normalize/modern-normalize.css';
 import { Layout } from './components';
 import { PrivateRoute } from './routes/PrivateRoute.jsx';
 import { PublicRoute } from './routes/PublicRoute.jsx';
-import { useAppDispatch, useAppSelector, useLocalStorage } from './hooks/';
-import {
-  fetchUserThunk,
-  refreshThunk,
-  selectIsLoggedIn,
-  selectRefreshToken,
-} from './redux';
 
 const WelcomePage = lazy(() => import('pages/WelcomePage/WelcomePage'));
 const AuthPage = lazy(() => import('pages/AuthPage/AuthPage.jsx'));
@@ -24,48 +16,6 @@ const ScreensPage = lazy(() => import('./pages/ScreensPage/ScreensPage.jsx'));
 const NotFoundPage = lazy(() => import('pages/NotFoundPage/NotFoundPage'));
 
 function App() {
-  const isLoggedIn = useAppSelector(selectIsLoggedIn);
-  const rToken = useAppSelector(selectRefreshToken);
-  const dispatch = useAppDispatch();
-  const [activeBoardId] = useLocalStorage('activeBoardId', '');
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (isLoggedIn) return;
-    const controller = new AbortController();
-
-    dispatch(
-      refreshThunk({
-        data: { refreshToken: rToken },
-        signal: controller.signal,
-      })
-    )
-      .unwrap()
-      .then(() => {
-        dispatch(fetchUserThunk({ boardId: activeBoardId }))
-          .unwrap()
-          .then(({ result }) => {
-            const activeBoard = result.boards.find((board) => board.active);
-            if (activeBoard) {
-              navigate(`/home/${activeBoard.title}`);
-            }
-          })
-          .catch((error) => {
-            toast.error(error.message);
-          });
-      })
-      .catch((error) => {
-        if (error === 'canceled')
-          toast.warn(
-            'Please wait, we use free servers for the back end and they need time to wake up'
-          );
-        toast.error(error.message);
-      });
-    setTimeout(() => {
-      controller.abort();
-    }, 5000);
-  }, [dispatch, navigate, isLoggedIn, rToken, activeBoardId]);
-
   return (
     <Routes>
       <Route path="/" element={<Layout />}>
